fix: consolidate feed to home route
