+++ README.md
@@ -1,7 +1,7 @@
 # RelayReel
 Nostr based short videos
 
-Visit the home feed at `/feed` or `/home`.
+Visit the home feed at `/home`.
 
 ## Development
 Start the Next.js dev server on all network interfaces:
